@huoyaoyuan 2022-03-15
It will be interesting to benchmark the cost of `Stopwatch` itself and reduce it.