@AlgorithmsAreCool 2022-03-15
I wonder if a new struct based 'timestamp' API would be of benefit for this case also.