@teo-tsirpanis 2022-03-15
@AlgorithmsAreCool a `ValueStopwatch` type in the BCL has been superseded by #65858.