@kevin-montrose 2022-03-15
@AlgorithmsAreCool The `Stopwatch` interface on a mutable struct is kind of "two shotguns duct-taped to nun-chucks" in terms of usability, so it'd have to be a pretty different API.  I suspect the dissonance of having two different "Stopwatches" with different interface in the BCL would be a strong point against it.  That said, it is nice to have a stronger type than just `long` to represent some "running timer thing-y" - which is part of why I went that route in the original project.