@kevin-montrose 2022-03-16
Looking over this once more I feel like the last bit of performance to eek out would be some sort of specialization around `s_tickFrequency`.  It's a constant (`0.001` I think) on unix-likes, and _often_ `1.0` on windows.  The JIT could figure it out... but doesn't seem to.

There's probably some incantation involving `static readonly` inlining - at least for the windows case - but it is unknown to me.