@danmoseley 2022-03-16
Does anyone have data suggesting that it's common to have large numbers of Stopwatch objects? We did "clever" things to make generic Dictionary smaller, but in that case we know there's lots of heaps with many of them.

Having said that I don't feel particularly strongly. @stephentoub as you touched this last (https://github.com/dotnet/runtime/issues/65858) do you have any opinion? Including about losing the negative range?