@stephentoub 2022-03-16
I appreciate the motivation to make things smaller, but if you have so many of these that there's meaningful benefit to shrinking the 40 byte object to 24 bytes, it'd be even better to just use GetTimestamp/GetElapsedTime with either a local (no allocation) or putting the long on whatever object is already carrying around the relevant state and increasing its size by only 8 bytes, or if it was carrying the Stopwatch reference, just replacing that reference with the long, in all cases avoiding the extra object altogether. If shrinking Stopwatch were "free", great, but this is increasing complexity and carries compat risk. At which point, while I applaud the effort, I don't think it's worth it.