@AlgorithmsAreCool 2022-03-16
> The Stopwatch interface on a mutable struct...

I wouldn't suggest a mutable struct stopwatch, but something that wrapped the frequency and timestamps. But it looks like #65858 was already basically what I was looking for.