@kevin-montrose 2022-03-16
Yeah, I'm curious at the prevalence of "lots of Stopwatches".  I was a tad shocked to find as many of them as I did (on the order of 1% of allocations, with a bunch of caveats since I got that through perfview's allocation sampling) when I went digging in this application.  They arose "naturally" with different code owners introducing them separately across a pipeline, so each "request" ended up with a few different Stopwatches.  All used for metadata too, which made them all feel extra wasteful.

I'm not sure how you'd establish if it's a common problem.  I have observed that "ValueStopwatch" tends to arise in a different codebases, _presumably_ because lots of Stopwatches were problems - but that's anecdotal at best.