@NickCraver 2022-03-16
For what's it's worth, I've also pulled a copy of ValueStopwatch into about a dozen codebases now because they had substantial amounts of allocations (>1%) that were purely Stopwatch. I find this very common in production services, often getting times for logging of pieces of the pipe, so every unit of work (e.g. request) tends to grow many. I don't know if this is the correct path or not, but I do agree Stopwatch is a legitimate problem of size I keep finding.

I was personally hoping ValueStopwatch was going to land in the framework from the earlier issue, because structure and intent (vs. a raw long) is much easier to explain, migrate to, and maintain for future developers. I think the long we've exposed of course works well if you understand it, but that has a cost for every consumer and maintainer downstream. I'd rather see any complexity internalized to something every user downstream doesn't have to deal with in whatever form.

Since the base problem is "there are enough of these to matter", converting to longs everywhere is harder to sell and I can't disagree with pushback on it - for others maintaining the code maybe it's not worth it and we continue to eat the allocations. Where does that discussion usually lead? Another copy of ValueStopwatch :)