@kevin-montrose 2022-03-17
@stephentoub 

> Anyone who wants to give the long a name and for whom a variable or field name is insufficient can trivially do so (same goes for any other unit-less type), e.g.

That works when you have a local `Stopwatch` and want to (effectively) stop and measure in "nearby" code.  It's not a replacement for the case where the `Stopwatch` is stopped "far" from where it's allocated (like in a field, or down in a called method).

Imagine something like:
```csharp
public class TrackWork
{
   private readonly Stopwatch workTracker = new Stopwatch();

   public TimeSpan SpentWorking => workTracker.Elapsed;

   public void OnWorkStarted()
   => workTracker.Start();

   public void OnWorkCompleted()
   => workTracker.Stop();
}
```

Nor does it handle cases where you repeatedly start and stop a `Stopwatch`, like if this `TrackWork` could have each `OnWorkXXX` called multiple times.

For those cases, you need multiple longs or to do some bit packing.

I do agree that that struct is trivial code to introduce, but it also only handles the simplest case.

---

@NickCraver Yeah, looking around Github I do see more examples of hacking around `Stopwatch` allocations.  I should probably build a list, as motivation for the proposal.  Presumably a more svelte `Stopwatch` would have saved some number of those folks from having to care about allocations, and improve things for some folks who should have been caring but don't know it.

I also need to spend some more time with the windows and unix apis and reason out the compatibility risk - I'm starting to suspect this scheme doesn't actually give up a bit (or, equivalently, the current implementation is not well behaved around rollover), but need to properly confirm.

Assuming those two things, I think there's more of an argument to be made for a change like this - less risk, more demonstrated upside.

All that  belongs in an Issue though, not a draft PR.  Kinda surprised a draft still triggered a bunch of bots honestly - not that I resent the feedback, I just hadn't expected it.